requirements.txt
asgiref==3.10.0
Django==4.1.7
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2